azure-func-appinsight/requirements.txt
azure-functions
requests>=2.23.0